Fix for non-extended addressing. Also fail on warnings
git-svn-id: svn://kore/repository/history/trunk@810 d6098cfc-5bd4-0310-8f2b-90d8aa983419

+++ utils/h16-plt2ps.cc
@@ -621,7 +621,7 @@ static bool parse_double(const std::string &str, double &v)
   const char *nptr = str.c_str();
   char *endptr;
   v = strtod(nptr, &endptr);
-  return ((endptr-nptr) == str.size());
+  return ((endptr-nptr) == ((signed long) str.size()));
 }
 
 static bool ascii_file;
